feature-envy: Fix static property and method handling
Previously, feature-envy treated references to static properties
and methods as references to a foreign class, thus making methods
fail the test when they should actually pass.

Instead, we now check if the accessed property / method is a
static member of the calling method's class and treat it like
a regular property / method.

Also added test cases to check for both passing / failing in the
static property / method contexts.